Remove unused subjects.js
