+++ openlibrary/plugins/openlibrary/js/index.js
@@ -15,8 +15,7 @@ import { enumerate, htmlquote, websafe, foreach, join, len, range } from './jsde
 import initAnalytics from './ol.analytics';
 import init from './ol.js';
 import * as Browser from './Browser';
-import { commify } from './python';
-import { Subject, urlencode, slice } from './subjects';
+import { commify, urlencode, slice } from './python';
 import Template from './template.js';
 // Add $.fn.focusNextInputField
 import { truncate, cond } from './utils';
@@ -45,7 +44,6 @@ window.ungettext = ungettext;
 window.uggettext = ugettext;
 
 window.Browser = Browser;
-window.Subject = Subject;
 window.Template = Template;
 
 // Extend existing prototypes
